use local instances of expandOnLoad
