+++ website/static/js/nodeSelectTreebeard.js
@@ -6,6 +6,60 @@ var ko = require('knockout');
 var Treebeard = require('treebeard');
 var projectSettingsTreebeardBase = require('js/projectSettingsTreebeardBase');
 
+function expandOnLoad() {
+    var tb = this;  // jshint ignore: line
+    for (var i = 0; i < tb.treeData.children.length; i++) {
+        var parent = tb.treeData.children[i];
+        tb.updateFolder(null, parent);
+        expandChildren(tb, parent.children);
+    }
+}
+
+function expandChildren(tb, children) {
+    var openParent = false;
+    for (var i = 0; i < children.length; i++) {
+        var child = children[i];
+        var parent = children[i].parent();
+        if (child.data.kind === 'event' && child.data.event.notificationType !== 'adopt_parent') {
+            openParent = true;
+        }
+        if (child.children.length > 0) {
+            expandChildren(tb, child.children);
+        }
+    }
+    if (openParent) {
+        openAncestors(tb, children[0]);
+    }
+}
+
+function openAncestors (tb, item) {
+    var parent = item.parent();
+    if(parent && parent.id > 0) {
+        tb.updateFolder(null, parent);
+        openAncestors(tb, parent);
+    }
+}
+
+function subscribe(item, notification_type) {
+    var id = item.parent().data.node.id;
+    var event = item.data.event.title;
+    var payload = {
+        'id': id,
+        'event': event,
+        'notification_type': notification_type
+    };
+    $osf.postJSON(
+        '/api/v1/subscriptions/',
+        payload
+    ).done(function(){
+        //'notfiy-success' is to override default class 'success' in treebeard
+        item.notify.update('Settings updated', 'notify-success', 1, 2000);
+        item.data.event.notificationType = notification_type;
+    }).fail(function() {
+        item.notify.update('Could not update settings', 'notify-danger', 1, 2000);
+    });
+}
+
 function NodeSelectTreebeard(divID, data, nodesState) {
     /**
      *  nodesState is a knockout variable that syncs the mithril checkbox list with information on the view.  The
@@ -78,7 +132,7 @@ function NodeSelectTreebeard(divID, data, nodesState) {
         }
     });
     var grid = new Treebeard(tbOptions);
-    projectSettingsTreebeardBase.expandOnLoad.call(grid.tbController);
+    expandOnLoad.call(grid.tbController);
 }
 module.exports = NodeSelectTreebeard;
