+++ website/static/js/nodesPrivacySettingsTreebeard.js
@@ -7,6 +7,37 @@ var Treebeard = require('treebeard');
 var $osf = require('js/osfHelpers');
 var projectSettingsTreebeardBase = require('js/projectSettingsTreebeardBase');
 
+function expandOnLoad() {
+    var tb = this;  // jshint ignore: line
+    for (var i = 0; i < tb.treeData.children.length; i++) {
+        var parent = tb.treeData.children[i];
+        tb.updateFolder(null, parent);
+        expandChildren(tb, parent.children);
+    }
+}
+
+function expandChildren(tb, children) {
+    var openParent = false;
+    for (var i = 0; i < children.length; i++) {
+        var child = children[i];
+        var parent = children[i].parent();
+        if (child.children.length > 0) {
+            expandChildren(tb, child.children);
+        }
+    }
+    if (openParent) {
+        openAncestors(tb, children[0]);
+    }
+}
+
+function openAncestors (tb, item) {
+    var parent = item.parent();
+    if(parent && parent.id > 0) {
+        tb.updateFolder(null, parent);
+        openAncestors(tb, parent);
+    }
+}
+
 function NodesPrivacyTreebeard(divID, data, nodesState, nodesOriginal) {
     /**
      * nodesChanged and nodesState are knockout variables.  nodesChanged will keep track of the nodes that have
@@ -81,7 +112,7 @@ function NodesPrivacyTreebeard(divID, data, nodesState, nodesOriginal) {
         }
     });
     var grid = new Treebeard(tbOptions);
-    projectSettingsTreebeardBase.expandOnLoad.call(grid.tbController);
+    expandOnLoad.call(grid.tbController);
 }
 module.exports = NodesPrivacyTreebeard;
